Add test_settings.
