+++ safe/gui/tools/wizard/test/test_wizard_util.py
@@ -5,14 +5,7 @@ import unittest
 from safe.test.utilities import get_qgis_app
 QGIS_APP, CANVAS, IFACE, PARENT = get_qgis_app()
 
-from PyQt4.QtCore import QSettings
-from safe.definitionsv4.constants import zero_default_value, RECENT, GLOBAL
-from safe.gui.tools.wizard.wizard_utils import (
-    get_question_text,
-    get_inasafe_default_value_qsetting
-)
-from utilities.settings import set_inasafe_default_value_qsetting, \
-    get_inasafe_default_value_qsetting
+from safe.gui.tools.wizard.wizard_utils import get_question_text
 
 __copyright__ = "Copyright 2016, The InaSAFE Project"
 __license__ = "GPL version 3"
@@ -28,27 +21,6 @@ class TestWizardUtil(unittest.TestCase):
         expected_text = '<b>MISSING CONSTANT: %s</b>' % constant
         text = get_question_text(constant)
         self.assertEqual(text, expected_text)
-
-    def test_inasafe_default_value_qsetting(self):
-        """Test for set and get inasafe_default_value_qsetting."""
-        qsetting = QSettings('InaSAFETest')
-        # Make sure it's empty
-        qsetting.clear()
-
-        female_ratio_key = 'female_ratio'
-        real_value = get_inasafe_default_value_qsetting(
-            qsetting, RECENT, female_ratio_key)
-        self.assertEqual(zero_default_value, real_value)
-
-        female_ratio_value = 0.8
-        set_inasafe_default_value_qsetting(
-            qsetting, RECENT, female_ratio_key, female_ratio_value)
-        real_value = get_inasafe_default_value_qsetting(
-            qsetting, RECENT, female_ratio_key)
-        self.assertEqual(female_ratio_value, real_value)
-
-        # Make sure it's empty
-        qsetting.clear()
 
 
 if __name__ == '__main__':
